+++ src/validation/decorators/index.ts
@@ -18,6 +18,10 @@ export {
   IsObjectWithKeysMaxLength,
   isObjectWithKeysMaxLength,
 } from './is-object-with-keys-max-length.decorator.js';
+export {
+  IsObjectWithValuesSatisfying,
+  isObjectWithValuesSatisfying,
+} from './is-object-with-values-satisfying.js';
 export { RequiresPropertyNotToBeDefined } from './requires-property-not-to-be-defined.decorator.js';
 export { RequiresProperty } from './requires-property.decorator.js';
 export { ValidateNestedType } from './validate-nested-type.decorator.js';